Subclass instead of god class
